refactor: use GitHubRepo type across all components
